# static/swagger.yml
swagger: "2.0"
info:
  description: "This is a python application"
  version: "1.0.0"
  title: "Python Application"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost"
basePath: "/api"
tags:
- name: "article"
  description: "Articles"
- name: "user"
  description: "Users"
schemes:
- "https"
- "http"
paths:
  /user/{user_id}/articles:
    post:
      tags:
      - "article"
      summary: 'Create an article'
      description: ""
      parameters:
      - name: "user_id"
        in: "path"
        description: "user id"
        required: true
        type: "integer"
      - name: "body"
        in: "body"
        required: true
        schema:
          type: "object"
          properties:
            link:
              type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "object"
            properties:
              success:
                type: "boolean"
              result:
                type: "object"
                properties:
                  id:
                    type: "integer"
        "400":
          description: "Invalid user_id supplied"
        "404":
          description: "Article not found"
    get:
      tags:
      - "article"
      summary: 'Get articles by user id'
      description: ""
      parameters:
      - name: "user_id"
        in: "path"
        description: "user id"
        required: true
        type: "integer"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "object"
            properties:
              success: 
                type: "boolean"
              result:
                type: "array"
                items:
                  $ref: "#/definitions/Article"
        "400":
          description: "Invalid user_id supplied"
        "404":
          description: "Article not found"
          
  /user/{user_id}/articles/{article_id}:
    get:
      tags:
      - "article"
      summary: 'Get article by article id'
      description: ""
      parameters:
      - name: "user_id"
        in: "path"
        description: "user id"
        required: true
        type: "integer"
      - name: "article_id"
        in: "path"
        description: "article id"
        required: true
        type: "integer"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Article"
        "400":
          description: "Invalid article_id supplied"
        "404":
          description: "Article not found"
    put:
      tags:
      - "article"
      summary: 'Update article'
      description: ""
      parameters:
      - name: "user_id"
        in: "path"
        description: "user id"
        required: true
        type: "integer"
      - name: "article_id"
        in: "path"
        description: "article id"
        required: true
        type: "integer"
      - name: "body"
        in: "body"
        required: true
        schema:
          type: "object"
          properties:
            title:
              type: "string"
            link:
              type: "string"
            summary:
              type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "object"
            properties:
              success:
                type: "boolean"
              result:
                type: "object"
                properties:
                  id:
                    type: "integer"
        "400":
          description: "Invalid article_id supplied"
        "404":
          description: "Article not found"
    delete:
      tags:
      - "article"
      summary: 'Delete article'
      description: ""
      parameters:
      - name: "user_id"
        in: "path"
        description: "user id"
        required: true
        type: "integer"
      - name: "article_id"
        in: "path"
        description: "article id"
        required: true
        type: "integer"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "object"
            properties:
              success:
                type: "boolean"
              result:
                type: "object"
                properties:
                  id:
                    type: "integer"
        "400":
          description: "Invalid article_id supplied"
        "404":
          description: "Article not found"
          
  
  /user/{user_id}/articles/{article_id}/archive:
    put:
      tags:
      - "article"
      summary: 'Archive article'
      description: ""
      parameters:
      - name: "user_id"
        in: "path"
        description: "user id"
        required: true
        type: "integer"
      - name: "article_id"
        in: "path"
        description: "article id"
        required: true
        type: "integer"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "object"
            properties:
              success:
                type: "boolean"
              result:
                type: "object"
                properties:
                  id:
                    type: "integer"
        "400":
          description: "Invalid article_id supplied"
        "404":
          description: "Article not found"
          
  /user:
    post:
      tags:
      - "user"
      summary: "Create user"
      description: ""
      parameters:
      - name: "body"
        in: "body"
        required: true
        schema:
          type: "object"
          properties:
            user_name:
              type: "string"
            password:
              type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "object"
            properties:
              success:
                type: "boolean"
              result:
                type: "object"
                properties:
                  id:
                    type: "integer"
        "400":
          description: "Invalid parameters supplied"
  
  /user/login:
    get:
      tags:
      - "user"
      summary: "Logs user into the system"
      description: ""
      parameters:
      - name: "username"
        in: "query"
        description: "The user name for login"
        required: true
        type: "string"
      - name: "password"
        in: "query"
        description: "The password for login in clear text"
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "object"
            properties:
              success:
                type: "boolean"
        "400":
          description: "Invalid username/password supplied"
  /user/logout:
    get:
      tags:
      - "user"
      summary: "Logs out current logged in user session"
      description: ""
      parameters: []
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "object"
            properties:
              success:
                type: "boolean"
  /user/{userId}:
    get:
      tags:
      - "user"
      summary: "Get user by user name"
      description: ""
      parameters:
      - name: "userId"
        in: "path"
        description: "The name that needs to be fetched. Use user1 for testing. "
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/User"
        "400":
          description: "Invalid username supplied"
        "404":
          description: "User not found"
    delete:
      tags:
      - "user"
      summary: "Delete user"
      description: ""
      parameters:
      - name: "userId"
        in: "path"
        description: "The name that needs to be deleted"
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "object"
            properties:
              success:
                type: "boolean"
        "400":
          description: "Invalid username supplied"
        "404":
          description: "User not found"
definitions:
  Article:
    type: "object"
    properties:
      id:
        type: "integer"
      title:
        type: "string"
      content:
        type: "string"
      top_image:
        type: "string"
        description: "the link of top image"
      summary:
        type: "string"
      status:
        type: "string"
        enum:
        - "created"
        - "updated"
        - "archived"
      created_date:
        type: "string"
        format: "date-time"
      updated_date:
        type: "string"
        format: "date-time"
  User:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      username:
        type: "string"
      password:
        type: "string"
      status:
        type: "integer"
        format: "int32"
        description: "User Status"